+++ tests/agent_unittests/test_harvest_loop.py
@@ -6,8 +6,11 @@ from testing_support.fixtures import override_generic_settings
 
 from newrelic.core.application import Application
 from newrelic.core.data_collector import send_request, collector_url
-from newrelic.core.stats_engine import CustomMetrics
+from newrelic.core.stats_engine import CustomMetrics, SampledDataSet
 from newrelic.core.transaction_node import TransactionNode
+from newrelic.core.custom_event import create_custom_event
+from newrelic.core.error_node import ErrorNode
+from newrelic.core.function_node import FunctionNode
 
 from newrelic.network.exceptions import RetryDataForRequest
 
@@ -15,7 +18,43 @@ settings = global_settings()
 
 
 @pytest.fixture(scope='module')
-def transaction_node():
+def transaction_node(request):
+    default_capacity = SampledDataSet().capacity
+    num_events = default_capacity + 1
+
+    custom_events = SampledDataSet(capacity=num_events)
+    for _ in range(num_events):
+        event = create_custom_event('Custom', {})
+        custom_events.add(event)
+
+    error = ErrorNode(
+            timestamp=0,
+            type='foo:bar',
+            message='oh no! your foo had a bar',
+            stack_trace='',
+            custom_params={},
+            file_name=None,
+            line_number=None,
+            source=None)
+
+    errors = tuple(error for _ in range(num_events))
+
+    function = FunctionNode(
+            group='Function',
+            name='foo',
+            children=(),
+            start_time=0,
+            end_time=1,
+            duration=1,
+            exclusive=1,
+            label=None,
+            params=None,
+            rollup=None,
+            is_async=True,
+            guid='GUID')
+
+    children = tuple(function for _ in range(num_events))
+
     node = TransactionNode(
             settings=finalize_application_settings({'agent_run_id': 1234567}),
             path='OtherTransaction/Function/main',
@@ -34,10 +73,10 @@ def transaction_node():
             response_time=0.1,
             duration=0.1,
             exclusive=0.1,
-            children=(),
-            errors=(),
+            children=children,
+            errors=errors,
             slow_sql=(),
-            custom_events=None,
+            custom_events=custom_events,
             apdex_t=0.5,
             suppress_apdex=False,
             custom_metrics=CustomMetrics(),
@@ -330,3 +369,27 @@ def test_ca_bundle(collector_agent_registration, ca_bundle_path,
             preconnect()
 
     _test()
+
+
+@override_generic_settings(settings, {
+    'developer_mode': True,
+    'license_key': '**NOT A LICENSE KEY**',
+    'feature_flag': set(),
+    'distributed_tracing.enabled': True,
+    'error_collector.max_event_samples_stored': 1000,
+    'span_events.max_samples_stored': 1000,
+    'custom_insights_events.max_samples_stored': 1000,
+})
+def test_reservoir_sizes(transaction_node):
+    app = Application('Python Agent Test (Harvest Loop)')
+    app.connect_to_data_collector()
+
+    # Record a transaction with events
+    app.record_transaction(transaction_node)
+
+    # Test that the samples have been recorded
+    assert app._stats_engine.custom_events.num_samples == 101
+    assert app._stats_engine.error_events.num_samples == 101
+
+    # Add 1 for the root span
+    assert app._stats_engine.span_events.num_samples == 102